sidebar修改

+++ web/service/src/main/java/service/com/emenu/service/stock/impl/SpecificationsServiceImpl.java
@@ -2,11 +2,14 @@ package com.emenu.service.stock.impl;
 
 import com.emenu.common.entity.dish.Unit;
 import com.emenu.common.entity.stock.Specifications;
+import com.emenu.common.entity.stock.StockItem;
 import com.emenu.common.exception.EmenuException;
 import com.emenu.mapper.stock.SpecificationsMapper;
 import com.emenu.service.dish.UnitService;
 import com.emenu.service.stock.SpecificationsService;
+import com.emenu.service.stock.StockItemService;
 import com.pandawork.core.common.exception.SSException;
+import com.pandawork.core.common.log.Log;
 import com.pandawork.core.common.log.LogClerk;
 import com.pandawork.core.common.util.Assert;
 import com.pandawork.core.framework.dao.CommonDao;
@@ -34,6 +37,9 @@ public class SpecificationsServiceImpl implements SpecificationsService {
     @Autowired
     private UnitService unitService;
 
+    @Autowired
+    private StockItemService stockItemService;
+
     @Override
     @Transactional(propagation = Propagation.REQUIRED, rollbackFor = {SSException.class, Exception.class, RuntimeException.class})
     public void add(Specifications specifications) throws SSException {
@@ -52,10 +58,12 @@ public class SpecificationsServiceImpl implements SpecificationsService {
     @Transactional(propagation = Propagation.REQUIRED, rollbackFor = {SSException.class, Exception.class, RuntimeException.class})
     public void deleteById(int id) throws SSException {
         if (id <= 0) return;
-
         try {
+            Specifications specifications = queryById(id);
+            if(!checkBeforeUpdateOrDelete(specifications)){
+                throw SSException.get(EmenuException.SpecificationIsUsed);
+            }
             commonDao.deleteById(Specifications.class, id);
-
         } catch (Exception e) {
             LogClerk.errLog.error(e);
             throw SSException.get(EmenuException.DeleteSpecificationsFail, e);
@@ -64,25 +72,38 @@ public class SpecificationsServiceImpl implements SpecificationsService {
 
     @Override
     @Transactional(propagation = Propagation.REQUIRED, rollbackFor = {SSException.class, Exception.class, RuntimeException.class})
-    public void update(Integer id, Specifications specifications) throws SSException {
-        if (id <= 0) return;
-
+    public void update(Specifications specifications) throws SSException {
         try {
             if (Assert.isNull(specifications)) {
                 throw SSException.get(EmenuException.SpecificationsIsNull);
             }
-            specificationsMapper.update(id, specifications);
+            if(!checkBeforeUpdateOrDelete(specifications)){
+                throw SSException.get(EmenuException.SpecificationIsUsed);
+            }
+            commonDao.update(specifications);
         } catch (Exception e) {
             LogClerk.errLog.error(e);
             throw SSException.get(EmenuException.UpdateSpecificationsFail, e);
         }
     }
 
+    private boolean checkBeforeUpdateOrDelete(Specifications specifications) throws SSException{
+            List<StockItem> stockItemList = stockItemService.listAll();
+            for(StockItem stockItem:stockItemList){
+                List<Integer> specificationList = stockItemService.stringTolist(stockItem.getSpecifications());
+                for(Integer specificationId:specificationList){
+                    if(specificationId == specifications.getId()){
+                        return false;
+                    }
+                }
+            }
+            return true;
+    }
+
     @Override
     public Specifications queryById(int id) throws SSException {
         Specifications specifications = new Specifications();
         if (id <= 0) return null;
-
         try {
             specifications = commonDao.queryById(specifications.getClass(), id);
             return specifications;
